:bug: fix LazyPhoto sporatically fails to load image in React 16.
I instrumented the code to figure this out, but it appears that if the img tag src is change to a value, changed to another value and changed back to the first value (perhaps any previous value) the onLoad never gets called or some other case of onLoad not getting called.  Either way, kinda makes this way of loading blank image first and then loading the real image doesn't really work.  Will consider updating to a more React way where in you render blank image when the image is not visable in the parent overflow